==> recsys2022/code/data_process/2_data_check.ipynb <==

valid:
check value correct
check list lengh

import pandas as pd
import os
import numpy as np

data_path = "/home/vmagent/app/transformer4rec_recsys2022"

%%time
train_bak = pd.read_parquet(os.path.join(data_path, "data/data_feat/bak/train.parquet"))
train = pd.read_parquet(os.path.join(data_path, "data/data_feat/train.parquet"))

train_bak = train_bak.drop((train_bak[train_bak["session_id"]==2132889]).index).reset_index(drop=True)
train_bak = train_bak.drop((train_bak[train_bak["session_id"]==3805637]).index).reset_index(drop=True)
train_bak = train_bak.drop((train_bak[train_bak["session_id"]==3823880]).index).reset_index(drop=True)

cols = ['item_id-list', 'f_1-list', 'f_2-list', 'f_3-list',
       'f_4-list', 'f_5-list', 'f_6-list', 'f_7-list', 'f_8-list', 'f_9-list',
       'f_10-list', 'f_11-list', 'f_12-list', 'f_13-list', 'f_14-list',
       'f_15-list', 'f_16-list', 'f_17-list', 'f_18-list', 'f_19-list',
       'f_20-list', 'f_21-list', 'f_22-list', 'f_23-list', 'f_24-list',
       'f_25-list', 'f_26-list', 'f_27-list', 'f_28-list', 'f_29-list',
       'f_30-list', 'f_31-list', 'f_32-list', 'f_33-list', 'f_34-list',
       'f_35-list', 'f_36-list', 'f_37-list', 'f_38-list', 'f_39-list',
       'f_40-list', 'f_41-list', 'f_42-list', 'f_43-list', 'f_44-list',
       'f_45-list', 'f_46-list', 'f_47-list', 'f_48-list', 'f_49-list',
       'f_50-list', 'f_51-list', 'f_52-list', 'f_53-list', 'f_54-list',
       'f_55-list', 'f_56-list', 'f_57-list', 'f_58-list', 'f_59-list',
       'f_60-list', 'f_61-list', 'f_62-list', 'f_63-list', 'f_64-list',
       'f_65-list', 'f_66-list', 'f_67-list', 'f_68-list', 'f_69-list',
       'f_70-list', 'f_71-list', 'f_72-list', 'f_73-list']

diff = train_bak - train

valid = pd.read_parquet(os.path.join(data_path, "data/data_feat/valid.parquet"))

sessions = pd.read_csv(os.path.join(data_path, "data/origin/valid_sessions_new.csv"))
purchases = pd.read_csv(os.path.join(data_path, "data/origin/valid_purchases_new.csv"))
wf = pd.read_parquet(os.path.join(data_path, "data/data_list/test.parquet"))[["session_id","wf"]]

train[train["session_id"]==4439999]["wf"].values

purchases

sessions[sessions["session_id"]==4439949].sort_values(["date"])

item_features = pd.read_csv(os.path.join(data_path, "data/origin/item_features.csv"))

item_features[(item_features["item_id"]==2171)&(item_features["feature_category_id"]==68)]

valid

train.columns

cols = ['item_id-list', 'f_1-list', 'f_2-list', 'f_3-list',
       'f_4-list', 'f_5-list', 'f_6-list', 'f_7-list', 'f_8-list', 'f_9-list',
       'f_10-list', 'f_11-list', 'f_12-list', 'f_13-list', 'f_14-list',
       'f_15-list', 'f_16-list', 'f_17-list', 'f_18-list', 'f_19-list',
       'f_20-list', 'f_21-list', 'f_22-list', 'f_23-list', 'f_24-list',
       'f_25-list', 'f_26-list', 'f_27-list', 'f_28-list', 'f_29-list',
       'f_30-list', 'f_31-list', 'f_32-list', 'f_33-list', 'f_34-list',
       'f_35-list', 'f_36-list', 'f_37-list', 'f_38-list', 'f_39-list',
       'f_40-list', 'f_41-list', 'f_42-list', 'f_43-list', 'f_44-list',
       'f_45-list', 'f_46-list', 'f_47-list', 'f_48-list', 'f_49-list',
       'f_50-list', 'f_51-list', 'f_52-list', 'f_53-list', 'f_54-list',
       'f_55-list', 'f_56-list', 'f_57-list', 'f_58-list', 'f_59-list',
       'f_60-list', 'f_61-list', 'f_62-list', 'f_63-list', 'f_64-list',
       'f_65-list', 'f_66-list', 'f_67-list', 'f_68-list', 'f_69-list',
       'f_70-list', 'f_71-list', 'f_72-list', 'f_73-list']

%%time
for col in cols:
    print(col)
    valid[f"{col}-len"] = valid[col].apply(lambda x: len(x))

valid

lens = ['item_id-list-len', 'f_1-list-len', 'f_2-list-len', 'f_3-list-len', 'f_4-list-len', 'f_5-list-len', 'f_6-list-len', 'f_7-list-len', 'f_8-list-len', 'f_9-list-len', 'f_10-list-len', 'f_11-list-len', 'f_12-list-len', 'f_13-list-len', 'f_14-list-len', 'f_15-list-len', 'f_16-list-len', 'f_17-list-len', 'f_18-list-len', 'f_19-list-len', 'f_20-list-len', 'f_21-list-len', 'f_22-list-len', 'f_23-list-len', 'f_24-list-len', 'f_25-list-len', 'f_26-list-len', 'f_27-list-len', 'f_28-list-len', 'f_29-list-len', 'f_30-list-len', 'f_31-list-len', 'f_32-list-len', 'f_33-list-len', 'f_34-list-len', 'f_35-list-len', 'f_36-list-len', 'f_37-list-len', 'f_38-list-len', 'f_39-list-len', 'f_40-list-len', 'f_41-list-len', 'f_42-list-len', 'f_43-list-len', 'f_44-list-len', 'f_45-list-len', 'f_46-list-len', 'f_47-list-len', 'f_48-list-len', 'f_49-list-len', 'f_50-list-len', 'f_51-list-len', 'f_52-list-len', 'f_53-list-len', 'f_54-list-len', 'f_55-list-len', 'f_56-list-len', 'f_57-list-len', 'f_58-list-len', 'f_59-list-len', 'f_60-list-len', 'f_61-list-len', 'f_62-list-len', 'f_63-list-len', 'f_64-list-len', 'f_65-list-len', 'f_66-list-len', 'f_67-list-len', 'f_68-list-len', 'f_69-list-len', 'f_70-list-len', 'f_71-list-len', 'f_72-list-len', 'f_73-list-len']

def average_len(df):
    sets = set([df[col] for col in lens])
    return sets

%%time
valid["len"] = valid.apply(average_len,axis=1)

valid["len-num"] = valid["len"].apply(lambda x: len(x))
valid[valid["len-num"]!=1]
