==> recsys2022/code/data_process/3_schema.ipynb <==
import pandas as pd
import os
import numpy as np

data_path = "/home/vmagent/app/transformer4rec_recsys2022/data/data_feat"

train = pd.read_parquet(os.path.join(data_path, "train.parquet"))
valid = pd.read_parquet(os.path.join(data_path, "valid.parquet"))

data = pd.concat([train,valid])

data

data["len"] = data["f_2-list"].apply(lambda x: len(x))
data["len"].min(), data["len"].max()

feature {
  name: "f2-list"
  value_count {
    min: 1
    max: 100
  }
  type: INT
  int_domain {
    name: "f2"
    min: 0
    max: 20
    is_categorical: true
  }
  annotation {
    tag: "categorical"
    tag: "list"
  }
}

cols = [f"f_{i}-list" for i in range(1,74)]

dict_list = []
for i in range(1,74):
    tmp_dict = {
        "name": f"f_{i}-list",
        "value_count": {"min":2, "max":101},
        "type": "INT",
        "int_domain":{
            "name":f"f_{i}",
            "min":0,
            "max":0,
            "is_categorical": True
        },
        "annotation": {
            "tag1": "categorical",
            "tag2": "list"
          }
    }
    dict_list.append(tmp_dict)

def get_min(x):
    min_v = np.array(x).min()
    return min_v
def get_max(x):
    max_v = np.array(x).max()
    return max_v

for i in range(1,74):
    data["min"]= data[f"f_{i}-list"].apply(get_min)
    data["max"]= data[f"f_{i}-list"].apply(get_max)
    min_v = data["min"].min()
    max_v = data["max"].max()
    print(i, min_v, max_v)
    dict_list[i-1]["int_domain"]["min"] = min_v
    dict_list[i-1]["int_domain"]["max"] = max_v

dict_list

dict_list[0]

def default_dump(obj):
    if isinstance(obj, (np.integer, np.floating, np.bool_)):
        return obj.item()
    elif instance(obj, np.ndarray):
        return obj.tolist()
    else:
        return obj

with open(os.path.join(data_path, "schema.txt"), "w") as f:
    for i in range(len(dict_list)):
        f.write("\nfeature {\n")
        f.write(json.dumps(dict_list[i],default=default_dump))
